+++ allennlp/data/dataset_readers/__init__.py
@@ -13,7 +13,6 @@ from allennlp.data.dataset_readers.conll2000 import Conll2000DatasetReader
 from allennlp.data.dataset_readers.ontonotes_ner import OntonotesNamedEntityRecognition
 from allennlp.data.dataset_readers.coreference_resolution import ConllCorefReader, WinobiasReader
 from allennlp.data.dataset_readers.dataset_reader import DatasetReader
-from allennlp.data.dataset_readers.elmo_lm import ElmoLMDatasetReader
 from allennlp.data.dataset_readers.event2mind import Event2MindDatasetReader
 from allennlp.data.dataset_readers.language_modeling import LanguageModelingReader
 from allennlp.data.dataset_readers.multiprocess_dataset_reader import MultiprocessDatasetReader
@@ -31,3 +30,4 @@ from allennlp.data.dataset_readers.quora_paraphrase import QuoraParaphraseDatase
 from allennlp.data.dataset_readers.semantic_parsing import (
         WikiTablesDatasetReader, AtisDatasetReader, NlvrDatasetReader, TemplateText2SqlDatasetReader)
 from allennlp.data.dataset_readers.semantic_parsing.quarel import QuarelDatasetReader
+from allennlp.data.dataset_readers.simple_language_modeling import SimpleLanguageModelingDatasetReader
